requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.1
python-dotenv==1.0.0
boto3==1.28.62
python-dateutil==2.8.2
alembic==1.12.1
